pyproject.toml
[project]
name = "field-classification-reworked"
version = "0.1.0"
description = "Qwen3VL-based image classification pipeline for satellite/aerial imagery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.47.0",
    "accelerate>=0.25.0",
    "pillow>=10.0.0",
    "tifffile>=2023.0.0",
    "imagecodecs>=2023.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.4.0",
    "h5py>=3.10.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "joblib>=1.3.0",
    "pandas>=2.0.0",
    "xgboost>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
